pyproject.toml
[project]
name = "markdown2kindle"
version = "0.1.0"
description = "A tool to convert and send Markdown files to a Kindle device."
authors = [
    {name = "astrisdesign",email = "clayton@astrisdesign.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">3.10,<4"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "markdown2 (>=2.5.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
markdown2kindle = "markdown2kindle.main:main"
